requirements/prod.txt
pydantic==2.5.3
fastapi==0.111.0
python-decouple==3.8
simple-settings==1.2.0